requirements.txt.txt
discord.py
selenium
flask
webdriver-manager
